Move Wsp specify encryption method inside CDP4WspDatabaseAuthentication project.
